Added security check for admin routes

+++ server/routes/admin.js
@@ -69,49 +69,77 @@ router.post('/movie', [upload.single('image'), middleware.checkAuthentification]
 })
 
 // Delete a movie
-router.delete('/movie/:id', (req, res) => {
-    Movie.deleteOne({_id: req.params.id})
+router.delete('/movie/:id', middleware.checkAuthentification, (req, res) => {
+
+    if (!req.user) {
+        return res.status(400).json(null)
+    }
+
+    User.findById(req.user.userId)
         .then(data => {
-            res.status(200).json(data)
-        })
-        .catch((err)  =>  {
-            res.status(400).json({
-                "Status": 400,
-                "Message": "Couldn't delete movie"
-            })
-        })
-})
+            if (!data.isAdmin) return res.status(401).json({"illegalRequest": true})
 
-// Update a movie
-router.put('/movie/:id', upload.single('image'), (req, res) => {
-
-    Movie.findById({_id: req.params.id})
-        .then(data => {
-            
-            data.title = req.body.title;
-            data.shortDescription = req.body.shortDescription;
-            data.longDescription = req.body.longDescription;
-            data.price = req.body.price;
-            //Only update image if a new image is provided 
-            if(req.file) {
-                data.img = {
-                    data: fs.readFileSync(path.join(__dirname + "/imageUploads/" + req.file.filename)),
-                    contentType: 'image/png'
-                };
-            }
-            data.genre = req.body.genre;
-
-            data.save()
+            else {
+                Movie.deleteOne({_id: req.params.id})
                 .then(data => {
                     res.status(200).json(data)
                 })
-                .catch((err) => {
+                .catch((err)  =>  {
                     res.status(400).json({
                         "Status": 400,
-                        "Message": "Couldn't update movie"
+                        "Message": "Couldn't delete movie"
                     })
                 })
+            }
         })
+        .catch(err => console.log(err))
+})
+
+// Update a movie
+router.put('/movie/:id', [upload.single('image'), middleware.checkAuthentification], (req, res) => {
+
+    if (!req.user) {
+        return res.status(400).json(null)
+    }
+
+    User.findById(req.user.userId)
+    .then(data => {
+        if (!data.isAdmin) return res.status(401).json({"illegalRequest": true})
+
+        else {
+            Movie.findById({_id: req.params.id})
+                .then(data => {
+                    
+                    data.title = req.body.title;
+                    data.shortDescription = req.body.shortDescription;
+                    data.longDescription = req.body.longDescription;
+                    data.price = req.body.price;
+                    //Only update image if a new image is provided 
+                    if(req.file) {
+                        data.img = {
+                            data: fs.readFileSync(path.join(__dirname + "/imageUploads/" + req.file.filename)),
+                            contentType: 'image/png'
+                        };
+                    }
+                    data.genre = req.body.genre;
+
+                    data.save()
+                        .then(data => {
+                            res.status(200).json(data)
+                        })
+                        .catch((err) => {
+                            res.status(400).json({
+                                "Status": 400,
+                                "Message": "Couldn't update movie"
+                            })
+                        })
+                })
+        }
+
+    })
+    .catch(err => console.log(err))
+
+    
 })
 
 module.exports = router;
